# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: UN ProjectAdvisor API
  description: |
    Knowledge Graph-powered AI system for legacy code documentation and institutional knowledge preservation.

    ## Architecture
    - **Graph Database**: Memgraph/Neo4j for knowledge storage
    - **Vector Store**: Qdrant for semantic search
    - **LLM Integration**: Ollama, Gemini, Anthropic for entity extraction
    - **Runtime Engine**: GXE DAG execution with AOPEG plugins
    - **Real-time**: SSE streaming and WebSocket support

    ## Authentication
    Currently no authentication required (internal use).

    ## Rate Limits
    No enforced rate limits. Batch endpoints have per-request limits.

    ## API Groups
    | Group | Endpoints | Description |
    |-------|-----------|-------------|
    | Query | 12 | Natural language query with hybrid retrieval |
    | Patterns | 11 | Entity & relation pattern management |
    | Graph-RAG | 13 | Knowledge graph CRUD and GNN-RAG |
    | Ingestion | 6 | Document upload and extraction pipeline |
    | Connectors | 12 | Source connector management |
    | Jobs | 9 | Background job queue management |
    | Visualization | 7 | Graph visualization data |
    | Dashboard | 10 | Metrics, analytics, activity |
    | Export | 4 | Graph export (6 formats) |
    | Reports | 8 | Report generation |
    | Runtime | 10 | GXE DAG execution engine |
    | AOPEG | 22 | AOPEG graph management |
    | AI Agent | 16 | AI Graph Builder agent |
...
  version: 1.0.0
  contact:
...
servers:
...
paths:
...
  /patterns/entity:
    get:
      tags: [Patterns]
      summary: List entity patterns
      parameters:
        - name: domain
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Patterns
    post:
      tags: [Patterns]
      summary: Create entity pattern
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityPatternInput"
      responses:
        "201":
          description: Created
  /patterns/entity/{id}:
    get:
      tags: [Patterns]
      summary: Get pattern
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pattern
        "404":
          description: Not found
    put:
      tags: [Patterns]
      summary: Update pattern
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityPatternInput"
      responses:
        "200":
          description: Updated
    delete:
      tags: [Patterns]
      summary: Delete pattern
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /patterns/relation:
    get:
      tags: [Patterns]
      summary: List relation patterns
      responses:
        "200":
          description: Patterns
    post:
      tags: [Patterns]
      summary: Create relation pattern
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationPatternInput"
      responses:
        "201":
          description: Created
...
  /patterns/match:
    post:
      tags: [Patterns]
      summary: Match patterns against text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                domain:
                  type: string
      responses:
        "200":
          description: Matches
...
components:
  schemas:
...
    EntityPatternInput:
      type: object
      required: [name, entityType, namePatterns]
      properties:
        name:
          type: string
        entityType:
          type: string
        namePatterns:
          type: array
          items:
            type: string
        contextKeywords:
          type: array
          items:
            type: string
        domain:
          type: string
          default: general
        priority:
          type: integer
          default: 5
        confidence:
          type: number
          default: 0.8
    RelationPatternInput:
      type: object
      required: [name, relationType, verbPatterns]
      properties:
        name:
          type: string
        relationType:
          type: string
        verbPatterns:
          type: array
          items:
            type: string
        subjectTypes:
          type: array
          items:
            type: string
        objectTypes:
          type: array
          items:
            type: string
...